Use applier change end signal to emit diff
